Simplify: move run function into index.ts
